src/web_app/pyproject.toml
[tool.poetry]
name = "web_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
aiogram = "^3.4.1"
apscheduler = "^3.10.4"
asyncpg = "^0.29.0"
environs = "^10.3.0"
fastapi = "^0.110.1"
jinja2 = "^3.1.3"
psycopg2 = "^2.9.9"
python-dateutil = "^2.9.0.post0"
setuptools = "^69.2.0"
sqlalchemy = "^2.0.28"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
